+++ NEWS.md
@@ -9,6 +9,7 @@ rearrangements of Notcurses.
     is automatically enlarged to accommodate output at its right (no scrolling)
     or bottom (scrolling enabled) boundaries.
   * Added `notcurses_default_background()` and `notcurses_default_foreground()`.
+  * Added `nccell_load_ucs32()`.
 
 * 3.0.0 (2021-12-01) **"In the A"**
   * Made the ABI/API changes that have been planned/collected during 2.x